feat: introduce v2 refresh token algorithm

+++ internal/tokens/service.go
@@ -15,6 +15,7 @@ import (
 
 	"github.com/supabase/auth/internal/api/apierrors"
 	"github.com/supabase/auth/internal/conf"
+	"github.com/supabase/auth/internal/crypto"
 	"github.com/supabase/auth/internal/hooks/v0hooks"
 	"github.com/supabase/auth/internal/metering"
 	"github.com/supabase/auth/internal/models"
@@ -130,7 +131,7 @@ func (s *Service) RefreshTokenGrant(ctx context.Context, db *storage.Connection,
 	for retry && time.Since(retryStart).Seconds() < retryLoopDuration {
 		retry = false
 
-		user, token, session, err := models.FindUserWithRefreshToken(db, params.RefreshToken, false)
+		user, anyToken, session, err := models.FindUserWithRefreshToken(db, config.Security.DBEncryption, params.RefreshToken, false)
 		if err != nil {
 			if models.IsNotFoundError(err) {
 				return nil, apierrors.NewBadRequestError(apierrors.ErrorCodeRefreshTokenNotFound, "Invalid Refresh Token: Refresh Token Not Found")
@@ -143,9 +144,11 @@ func (s *Service) RefreshTokenGrant(ctx context.Context, db *storage.Connection,
 		}
 
 		if session == nil {
-			// a refresh token won't have a session if it's created prior to the sessions table introduced
-			if err := db.Destroy(token); err != nil {
-				return nil, apierrors.NewInternalServerError("Error deleting refresh token with missing session").WithInternalError(err)
+			if token, ok := anyToken.(*models.RefreshToken); ok {
+				// a refresh token won't have a session if it's created prior to the sessions table introduced
+				if err := db.Destroy(token); err != nil {
+					return nil, apierrors.NewInternalServerError("Error deleting refresh token with missing session").WithInternalError(err)
+				}
 			}
 			return nil, apierrors.NewBadRequestError(apierrors.ErrorCodeSessionNotFound, "Invalid Refresh Token: No Valid Session Found")
 		}
@@ -159,7 +162,12 @@ func (s *Service) RefreshTokenGrant(ctx context.Context, db *storage.Connection,
 			AllowLowAAL:       config.Sessions.AllowLowAAL,
 		}
 
-		result := session.CheckValidity(sessionValidityConfig, retryStart, &token.UpdatedAt, user.HighestPossibleAAL())
+		var refreshTokenTime *time.Time
+		if token, ok := anyToken.(*models.RefreshToken); ok {
+			refreshTokenTime = &token.UpdatedAt
+		}
+
+		result := session.CheckValidity(sessionValidityConfig, retryStart, refreshTokenTime, user.HighestPossibleAAL())
 
 		switch result {
 		case models.SessionValid:
@@ -187,7 +195,7 @@ func (s *Service) RefreshTokenGrant(ctx context.Context, db *storage.Connection,
 		var newTokenResponse *AccessTokenResponse
 
 		err = db.Transaction(func(tx *storage.Connection) error {
-			user, token, session, terr := models.FindUserWithRefreshToken(tx, params.RefreshToken, true /* forUpdate */)
+			user, anyToken, session, terr := models.FindUserWithRefreshToken(tx, config.Security.DBEncryption, params.RefreshToken, true /* forUpdate */)
 			if terr != nil {
 				if models.IsNotFoundError(terr) {
 					// because forUpdate was set, and the
@@ -266,7 +274,7 @@ func (s *Service) RefreshTokenGrant(ctx context.Context, db *storage.Connection,
 					// since token is not the refresh token
 					// of s, we can't use it's UpdatedAt
 					// time to compare!
-					if s.LastRefreshedAt(nil).After(session.LastRefreshedAt(&token.UpdatedAt)) {
+					if s.LastRefreshedAt(nil).After(session.LastRefreshedAt(refreshTokenTime)) {
 						// session is not the most
 						// recently active one
 						return apierrors.NewBadRequestError(apierrors.ErrorCodeSessionExpired, "Invalid Refresh Token: Session Expired (Revoked by Newer Login)")
@@ -279,60 +287,157 @@ func (s *Service) RefreshTokenGrant(ctx context.Context, db *storage.Connection,
 			// refresh token row and session are locked at this
 			// point, cannot be concurrently refreshed
 
-			var issuedToken *models.RefreshToken
+			var issuedToken string
 
-			if token.Revoked {
-				activeRefreshToken, terr := session.FindCurrentlyActiveRefreshToken(tx)
-				if terr != nil && !models.IsNotFoundError(terr) {
-					return apierrors.NewInternalServerError(terr.Error())
-				}
+			if token, ok := anyToken.(*models.RefreshToken); ok {
+				if token.Revoked {
+					activeRefreshToken, terr := session.FindCurrentlyActiveRefreshToken(tx)
+					if terr != nil && !models.IsNotFoundError(terr) {
+						return apierrors.NewInternalServerError(terr.Error())
+					}
 
-				if activeRefreshToken != nil && activeRefreshToken.Parent.String() == token.Token {
-					// Token was revoked, but it's the
-					// parent of the currently active one.
-					// This indicates that the client was
-					// not able to store the result when it
-					// refreshed token. This case is
-					// allowed, provided we return back the
-					// active refresh token instead of
-					// creating a new one.
-					issuedToken = activeRefreshToken
-				} else {
-					// For a revoked refresh token to be reused, it
-					// has to fall within the reuse interval.
-					reuseUntil := token.UpdatedAt.Add(
-						time.Second * time.Duration(config.Security.RefreshTokenReuseInterval))
+					if activeRefreshToken != nil && activeRefreshToken.Parent.String() == token.Token {
+						// Token was revoked, but it's the
+						// parent of the currently active one.
+						// This indicates that the client was
+						// not able to store the result when it
+						// refreshed token. This case is
+						// allowed, provided we return back the
+						// active refresh token instead of
+						// creating a new one.
+						issuedToken = activeRefreshToken.Token
+					} else {
+						// For a revoked refresh token to be reused, it
+						// has to fall within the reuse interval.
+						reuseUntil := token.UpdatedAt.Add(
+							time.Second * time.Duration(config.Security.RefreshTokenReuseInterval))
 
-					if s.now().After(reuseUntil) {
-						// not OK to reuse this token
-						if config.Security.RefreshTokenRotationEnabled {
-							// Revoke all tokens in token family
-							if err := models.RevokeTokenFamily(tx, token); err != nil {
-								return apierrors.NewInternalServerError(err.Error())
+						if s.now().After(reuseUntil) {
+							// not OK to reuse this token
+							if config.Security.RefreshTokenRotationEnabled {
+								// Revoke all tokens in token family
+								if err := models.RevokeTokenFamily(tx, token); err != nil {
+									return apierrors.NewInternalServerError(err.Error())
+								}
 							}
-						}
 
-						return storage.NewCommitWithError(apierrors.NewBadRequestError(apierrors.ErrorCodeRefreshTokenAlreadyUsed, "Invalid Refresh Token: Already Used").WithInternalMessage("Possible abuse attempt: %v", token.ID))
+							return storage.NewCommitWithError(apierrors.NewBadRequestError(apierrors.ErrorCodeRefreshTokenAlreadyUsed, "Invalid Refresh Token: Already Used").WithInternalMessage("Possible abuse attempt: %v", token.ID))
+						}
 					}
 				}
-			}
 
-			if terr := models.NewAuditLogEntry(config.AuditLog, r, tx, user, models.TokenRefreshedAction, "", nil); terr != nil {
-				return terr
-			}
-
-			if issuedToken == nil {
-				newToken, terr := models.GrantRefreshTokenSwap(config.AuditLog, r, tx, user, token)
-				if terr != nil {
+				if terr := models.NewAuditLogEntry(config.AuditLog, r, tx, user, models.TokenRefreshedAction, "", nil); terr != nil {
 					return terr
 				}
 
-				issuedToken = newToken
+				if issuedToken == "" {
+					newToken, terr := models.GrantRefreshTokenSwap(config.AuditLog, r, tx, user, token)
+					if terr != nil {
+						return terr
+					}
+
+					issuedToken = newToken.Token
+				}
+			} else if token, ok := anyToken.(*crypto.RefreshToken); ok {
+				signingKey, _, kerr := session.GetRefreshTokenHmacKey(config.Security.DBEncryption)
+				if kerr != nil {
+					return apierrors.NewInternalServerError("failed to load session from database").WithInternalError(terr)
+				}
+
+				counterDifference := *session.RefreshTokenCounter - token.Counter
+
+				if counterDifference < 0 {
+					// refresh token was not issued by this server
+				} else if counterDifference == 0 || config.Security.RefreshTokenAllowReuse {
+					// normal refresh token use
+					counter := *session.RefreshTokenCounter + 1
+					session.RefreshTokenCounter = &counter
+
+					issuedToken = (&crypto.RefreshToken{
+						Version:   0,
+						SessionID: session.ID,
+						Counter:   *session.RefreshTokenCounter,
+					}).Encode(signingKey)
+				} else if counterDifference > 0 {
+					// refresh token is being reused
+
+					// This is caused when the client has
+					// failed to receive or save the
+					// response from the last refresh token
+					// requests. This occurs more
+					// frequently than you can imagine, so
+					// it's an allowed reuse.
+					likelyNotSavedByClient := counterDifference == 1
+
+					// Concurrent refreshes occur when the
+					// client sends off multiple refresh
+					// token requests at once or close by.
+					// Often this happens when your browser
+					// remembers multiple tabs of the app,
+					// which were paused by it or by the OS
+					// (such as you quitting the browser)
+					// and then opening it back up. If the
+					// app uses SSR it is likely that the N
+					// open tabs will immediately send a
+					// request to the app's hosting server,
+					// which will attempt to concurrently
+					// refresh the session at once using
+					// refresh token.
+					likelyConcurrentRefreshes := retryStart.Sub(session.LastRefreshedAt(nil)).Abs() < time.Duration(config.Security.RefreshTokenReuseInterval)*time.Second
+
+					reuseAllowed := likelyNotSavedByClient || likelyConcurrentRefreshes || config.Security.RefreshTokenAllowReuse
+
+					if reuseAllowed {
+						// When reuse is allowed, we do
+						// not increment the counter.
+						// This allows all of the
+						// concurrent clients to
+						// synchronize their state
+						// within the refresh token
+						// reuse interval to the
+						// currently active refresh
+						// token.
+
+						issuedToken = (&crypto.RefreshToken{
+							Version:   0,
+							SessionID: session.ID,
+							Counter:   *session.RefreshTokenCounter,
+						}).Encode(signingKey)
+					} else if config.Security.RefreshTokenRotationEnabled {
+						// Reuse is not allowed, in
+						// which case the whole session
+						// must go preventing any
+						// client with any refresh and
+						// access token for this
+						// session from being used.
+
+						if terr := models.LogoutSession(tx, session.ID); terr != nil {
+							return apierrors.NewInternalServerError("destroying session after detected refresh token reuse failed").WithInternalError(terr)
+						}
+
+						return storage.NewCommitWithError(apierrors.NewBadRequestError(apierrors.ErrorCodeRefreshTokenAlreadyUsed, "Invalid Refresh Token: Already Used").WithInternalMessage("Refresh token behind current counter by %v, session %v is terminated due to refresh token reuse", counterDifference, session.ID.String()))
+					} else {
+						// Reuse is not allowed, but no
+						// refresh token rotation
+						// enabled. So only fail this
+						// request.
+
+						return storage.NewCommitWithError(apierrors.NewBadRequestError(apierrors.ErrorCodeRefreshTokenAlreadyUsed, "Invalid Refresh Token: Already Used").WithInternalMessage("Refresh token behind current counter by %v, session %v is not terminated", counterDifference, session.ID.String()))
+					}
+				}
+
+				if terr := session.UpdateOnlyRefreshToken(tx); terr != nil {
+					return apierrors.NewInternalServerError("failed saving session").WithInternalError(terr)
+				}
+
+				if terr := models.NewAuditLogEntry(config.AuditLog, r, tx, user, models.TokenRefreshedAction, "", nil); terr != nil {
+					return terr
+				}
 			}
 
 			tokenString, expiresAt, terr = s.GenerateAccessToken(r, tx, GenerateAccessTokenParams{
 				User:                 user,
-				SessionID:            issuedToken.SessionId,
+				SessionID:            &session.ID,
 				AuthenticationMethod: models.TokenRefresh,
 				ClientID:             sessionClientID,
 			})
@@ -370,7 +475,7 @@ func (s *Service) RefreshTokenGrant(ctx context.Context, db *storage.Connection,
 				TokenType:    "bearer",
 				ExpiresIn:    config.JWT.Exp,
 				ExpiresAt:    expiresAt,
-				RefreshToken: issuedToken.Token,
+				RefreshToken: issuedToken,
 				User:         user,
 			}
 
@@ -473,28 +578,60 @@ func (s *Service) IssueRefreshToken(r *http.Request, conn *storage.Connection, u
 
 	var tokenString string
 	var expiresAt int64
-	var refreshToken *models.RefreshToken
+	var refreshToken string
+	var sessionID uuid.UUID
 	var oAuthClientID *uuid.UUID
 
 	err := conn.Transaction(func(tx *storage.Connection) error {
 		var terr error
 
-		refreshToken, terr = models.GrantAuthenticatedUser(tx, user, grantParams)
-		if terr != nil {
-			return apierrors.NewInternalServerError("Database error granting user").WithInternalError(terr)
+		if config.Security.RefreshTokenAlgorithmVersion == 2 {
+			session, terr := models.NewSession(user.ID, grantParams.FactorID)
+			if terr != nil {
+				return apierrors.NewInternalServerError("Failed to create new session").WithInternalError(terr)
+			}
+
+			session.ApplyGrantParams(&grantParams)
+			if terr := session.SetupRefreshTokenData(config.Security.DBEncryption); terr != nil {
+				return apierrors.NewInternalServerError("Failed to setup refresh token data for session").WithInternalError(terr)
+			}
+
+			if terr := tx.Create(session); terr != nil {
+				return apierrors.NewInternalServerError("Database error creating new session").WithInternalError(terr)
+			}
+
+			signingKey, _, terr := session.GetRefreshTokenHmacKey(config.Security.DBEncryption)
+			if terr != nil {
+				return apierrors.NewInternalServerError("Failed to get session's refresh token key").WithInternalError(terr)
+			}
+
+			sessionID = session.ID
+			refreshToken = (&crypto.RefreshToken{
+				SessionID: session.ID,
+				Counter:   *session.RefreshTokenCounter,
+			}).Encode(signingKey)
+		} else {
+			rt, terr := models.GrantAuthenticatedUser(tx, user, grantParams)
+			if terr != nil {
+				return apierrors.NewInternalServerError("Database error granting user").WithInternalError(terr)
+			}
+
+			sessionID = *rt.SessionId
+			refreshToken = rt.Token
 		}
+
 		if grantParams.OAuthClientID != nil && *grantParams.OAuthClientID != uuid.Nil {
 			oAuthClientID = grantParams.OAuthClientID
 		}
 
-		terr = models.AddClaimToSession(tx, *refreshToken.SessionId, authenticationMethod)
+		terr = models.AddClaimToSession(tx, sessionID, authenticationMethod)
 		if terr != nil {
 			return terr
 		}
 
 		tokenString, expiresAt, terr = s.GenerateAccessToken(r, tx, GenerateAccessTokenParams{
 			User:                 user,
-			SessionID:            refreshToken.SessionId,
+			SessionID:            &sessionID,
 			AuthenticationMethod: authenticationMethod,
 			ClientID:             oAuthClientID,
 		})
@@ -516,7 +653,7 @@ func (s *Service) IssueRefreshToken(r *http.Request, conn *storage.Connection, u
 		TokenType:    "bearer",
 		ExpiresIn:    config.JWT.Exp,
 		ExpiresAt:    expiresAt,
-		RefreshToken: refreshToken.Token,
+		RefreshToken: refreshToken,
 		User:         user,
 	}, nil
 }
